requirements.txt
geopy
pyyaml
gpsd-py3
